+++ tests/integrations/basic/tests/ui_tests.rs
@@ -1,4 +1,4 @@
-use ui_test::{dependencies::DependencyBuilder, *};
+use ui_test::{custom_flags::rustfix::RustfixMode, dependencies::DependencyBuilder, *};
 
 fn main() -> ui_test::color_eyre::Result<()> {
     let path = "../../../target";
@@ -19,6 +19,10 @@ fn main() -> ui_test::color_eyre::Result<()> {
         .comment_defaults
         .base()
         .set_custom("dependencies", DependencyBuilder::default());
+    config
+        .comment_defaults
+        .base()
+        .set_custom("rustfix", RustfixMode::Everything);
 
     if let Ok(target) = std::env::var("UITEST_TEST_TARGET") {
         config.target = Some(target);