+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unrealeased]
+
+### Added
+
+### Fixed
+
+### Changed
+
+* Split up `Revisioned::mode` into `Revisioned::exit_status` and `Revisioned::require_annotations`
+* `Config::output_conflict_handling` is now `Error` instead of `Bless`
+
+### Removed
+
+
+## [0.23.0] - 2024-05-02
 
 ### Added